requirements.txt
Django==5.0
Faker==21.0.0
django-debug-toolbar==4.2.0
python-dotenv==1.0.0
django-rq==2.10.1
rq==1.15.1
psycopg2-binary==2.9.9
djangorestframework==3.14.0
Markdown==3.5.2
django-filter==23.5
